--- Ballistics-master/calibers.txt
@@ -1,4 +1,0 @@
-#7.62 M80
-833
-#7.62 M118
-790
