+++ calibers.txt
@@ -1,0 +1,4 @@
+#7.62 M80
+833
+#7.62 M118
+790
